requirements.txt
Django==1.4.3
MySQL-python==1.2.4b4
PIL==1.1.7
South==0.7.6
Whoosh==2.4.1
amqp==1.0.8
anyjson==0.3.3
argparse==1.2.1
billiard==2.7.3.21
celery==3.0.15
distribute==0.6.28
dj-database-url==0.2.1
django-activity-stream==0.4.4
django-debug-toolbar==0.9.4
django-haystack==1.2.7
django-model-utils==1.2.0
django-notification==1.0
django-notifications-hq==0.5.5
django-ratings==0.3.7
django-recommends==0.0.11
django-registration==0.8
django-relationships==0.3.2
django-social-auth==0.7.19
django-tagging==0.3.1
django-tastypie==0.9.12
easy-thumbnails
gunicorn==0.17.2
httplib2==0.7.7
jsonfield==0.9.5
kombu==2.5.6
mimeparse==0.1.3
oauth2==1.5.211
ordereddict==1.1
#-e git+https://github.com/harshulj/django-privatebeta.git@2b5ddb6ff3c8a8e795b72d42cdfe196e6266072e#egg=privatebeta-dev
psycopg2==2.4.6
python-dateutil==1.5
python-memcached==1.48
python-openid==2.2.5
wsgiref==0.1.2
